VoG Connect - Copy/requirements.txt
Flask==2.2.5
gunicorn==22.0.0
Flask-Cors==4.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0 